allie/backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
httpx
python-dotenv
google-cloud-firestore
google-cloud-storage
sentence-transformers==2.5.1
psycopg[binary]
numpy
youtube-transcript-api
